+++ packages/sn-filter-pane/src/components/FoldedListbox/FoldedListbox.tsx
@@ -2,14 +2,14 @@ import { stardust } from '@nebula.js/stardust';
 import {
   Grid, styled, Tooltip, Typography,
 } from '@mui/material';
-import React, { useRef } from 'react';
+import React from 'react';
 import { IListboxResource } from '../../hooks/types';
 import useFieldName from '../../hooks/use-field-name';
 import type { IStores } from '../../store';
-import { COLLAPSED_HEIGHT } from '../ListboxGrid/distribute-resources';
 import DrillDown from './drillDown';
 import SelectionSegmentsIndicator from './SelectionSegmentsIndicator';
 import KEYS from '../keys';
+import getSizes from './get-sizes';
 
 export interface FoldedListboxClickEvent {
   event: React.MouseEvent<HTMLDivElement>;
@@ -25,14 +25,16 @@ export interface FoldedListboxProps {
 interface StyledGridProps {
   constraints?: stardust.Constraints;
   stardustTheme?: stardust.Theme;
-  containerRef: React.RefObject<HTMLDivElement>;
   isInPopover: boolean;
+  height: number;
 }
 interface StyledDivProps {
   isInPopover?: boolean;
 }
 
 const POPOVER_PADDING = 2;
+
+const getListboxStyle = (path: string, prop: string, t?: stardust.Theme) => t?.getStyle('object.listBox', path, prop);
 
 const StyledDiv = styled('div', { shouldForwardProp: (p) => !['isInPopover'].includes(p as string) })<StyledDivProps>(
   ({ isInPopover }) => ({
@@ -46,29 +48,28 @@ const StyledDiv = styled('div', { shouldForwardProp: (p) => !['isInPopover'].inc
   }),
 );
 
-const StyledGrid = styled(Grid, { shouldForwardProp: (p) => !['constraints', 'stardustTheme', 'containerRef', 'isInPopover'].includes(p as string) })<StyledGridProps>(
+const StyledGrid = styled(Grid, { shouldForwardProp: (p) => !['constraints', 'stardustTheme', 'isInPopover', 'height'].includes(p as string) })<StyledGridProps>(
   ({
     constraints,
     stardustTheme,
-    containerRef,
     isInPopover,
+    height,
   }) => {
-    const containerWidth = containerRef?.current?.clientWidth ?? 0;
     const popoverPadding = isInPopover ? POPOVER_PADDING * 2 : 0;
     return {
       justifyContent: 'space-between',
       alignItems: 'flex-start',
       border: '1px solid #d9d9d9',
       borderRadius: '3px',
-      height: COLLAPSED_HEIGHT,
+      height,
       overflow: 'hidden',
       cursor: constraints?.active ? 'default' : 'pointer',
       ':hover': !constraints?.active && {
         border: '1px solid #595959',
       },
-      backgroundColor: stardustTheme?.getStyle('object', '', 'listBox.backgroundColor') ?? '#FFFFFF',
-      color: stardustTheme?.getStyle('object', '', 'listBox.title.main.color') ?? '#404040',
-      width: containerWidth - popoverPadding,
+      backgroundColor: getListboxStyle('', 'backgroundColor', stardustTheme) ?? '#FFFFFF',
+      color: getListboxStyle('title.main', 'color', stardustTheme) ?? '#404040',
+      width: isInPopover ? `calc(100% - ${2 * popoverPadding}px` : '100%',
       '&:focus:not(:hover)': {
         boxShadow: 'inset 0 0 0 2px #3F8AB3 !important',
       },
@@ -79,16 +80,26 @@ const StyledGrid = styled(Grid, { shouldForwardProp: (p) => !['constraints', 'st
   },
 );
 
+const Title = styled(Typography, { shouldForwardProp: (p) => !['stardustTheme'].includes(p as string) })<{ stardustTheme?: stardust.Theme }>(
+  ({ stardustTheme }) => ({
+    fontSize: '13px',
+    fontFamily: getListboxStyle('title.main', 'fontFamily', stardustTheme) ?? '"Source Sans Pro", sans-serif',
+    fontWeight: getListboxStyle('title.main', 'fontWeight', stardustTheme) ?? '700',
+  }),
+);
+
 export const FoldedListbox = ({
   resource, onClick, stores, isInPopover,
 }: FoldedListboxProps) => {
   const fieldName = useFieldName(resource.layout);
-  const containerRef = useRef<HTMLDivElement>(null);
   const {
-    constraints, stardustTheme, options, translator,
+    constraints, stardustTheme, options, translator, containerSize,
   } = stores.store.getState();
   const isRtl = options.direction === 'rtl';
   const isDrillDown = resource.layout.qListObject.qDimensionInfo.qGrouping === 'H';
+  const {
+    gridHeight, narrowSmall, narrowLarge,
+  } = getSizes(containerSize, isInPopover);
 
   const handleKeyDown = (event: React.KeyboardEvent) => {
     const target = event.target as HTMLElement;
@@ -110,35 +121,40 @@ export const FoldedListbox = ({
   };
 
   return (
-    <StyledDiv ref={containerRef} className="folded-listbox" onKeyDown={handleKeyDown} isInPopover={isInPopover}>
+    <StyledDiv className="folded-listbox" onKeyDown={handleKeyDown} isInPopover={isInPopover}>
       <StyledGrid
         constraints={constraints}
         stardustTheme={stardustTheme}
-        containerRef={containerRef}
         container
         direction='column'
         data-testid={`collapsed-title-${fieldName}`}
         isInPopover={!!isInPopover}
-        onClick={(event) => onClick({ event, resource })}>
-        <Grid container flexGrow={1} alignItems={'center'} sx={{ flexDirection: isRtl ? 'row-reverse' : 'row', flexWrap: 'nowrap' }} padding='0 8px'>
-          {isDrillDown
-            && <Tooltip title={translator?.get('Listbox.DrillDown')} enterDelay={2000}>
-              <div>
-                < DrillDown style={{ fontSize: '12px', padding: isRtl ? '0 0 0 6px' : '0 6px 0 0' }} />
-              </div>
+        onClick={(event) => onClick({ event, resource })}
+        height={gridHeight}
+      >
+        {!narrowSmall
+          && <Grid container flexGrow={1} alignItems={'center'} sx={{ flexDirection: isRtl ? 'row-reverse' : 'row', flexWrap: 'nowrap' }} padding='0 8px'>
+            {isDrillDown
+              && <Tooltip title={translator?.get('Listbox.DrillDown')} enterDelay={2000}>
+                <div>
+                  < DrillDown style={{ fontSize: '12px', padding: isRtl ? '0 0 0 6px' : '0 6px 0 0' }} />
+                </div>
+              </Tooltip>
+            }
+            <Tooltip title={fieldName} enterDelay={2000}>
+              <Title variant="subtitle2" stardustTheme={stardustTheme} noWrap lineHeight='normal'>
+                {fieldName}
+              </Title>
             </Tooltip>
-          }
-          <Tooltip title={fieldName} enterDelay={2000}>
-            <Typography variant="subtitle2" fontSize='13px' noWrap>
-              {fieldName}
-            </Typography>
-          </Tooltip>
-        </Grid>
-        <Grid item width='100%'>
-          <SelectionSegmentsIndicator
-            qDimensionInfo={resource.layout.qListObject.qDimensionInfo}
-          ></SelectionSegmentsIndicator>
-        </Grid>
+          </Grid>
+        }
+        {!narrowLarge
+          && <Grid item width='100%' height={narrowSmall ? '100%' : undefined}>
+            <SelectionSegmentsIndicator
+              qDimensionInfo={resource.layout.qListObject.qDimensionInfo} fullHeight={narrowSmall}
+            ></SelectionSegmentsIndicator>
+          </Grid>
+        }
       </StyledGrid>
     </StyledDiv>
   );