+++ packages/sn-filter-pane/src/components/ListboxContainer.tsx
@@ -1,14 +1,17 @@
 import React, { useEffect, useRef, useState } from 'react';
+import extend from 'extend';
 import { Box } from '@mui/material';
 import { stardust } from '@nebula.js/stardust';
 import { IListLayout } from '../hooks/types';
 import type { IStores } from '../store';
 import uid from '../utils/uid';
+import { IEnv } from '../types/types';
+import useDirectQuery from '../hooks/direct-query/use-direct-query';
 
 interface ListboxContainerProps {
   layout: IListLayout;
+  model: EngineAPI.IGenericObject;
   borderBottom?: boolean;
-  disableSearch?: boolean;
   handleActive?: (id: string, active: boolean) => void;
   stores: IStores;
   closePopover?: () => void;
@@ -16,7 +19,7 @@ interface ListboxContainerProps {
 }
 
 const ListboxContainer = ({
-  layout, borderBottom, disableSearch = false, handleActive, stores, closePopover, isPopover,
+  layout, model, borderBottom, handleActive, stores, closePopover, isPopover,
 }: ListboxContainerProps) => {
   const [listboxInstance, setListboxInstance] = useState<stardust.FieldInstance>();
   const elRef = useRef<HTMLElement>();
@@ -24,13 +27,19 @@ const ListboxContainer = ({
 
   const {
     embed,
-    model,
     constraints,
     options,
     renderTracker,
-    sense,
+    env,
     stardustTheme,
+    directQueryEnabled,
   } = stores.store.getState();
+
+  const { sense } = env as IEnv;
+
+  const dqOptionsOverrides = useDirectQuery({
+    directQueryEnabled, layout, listBoxModel: model, constraints,
+  });
 
   const showBorder = !sense || inSelection || stardustTheme?.getStyle('', '', '_cards');
 
@@ -51,14 +60,15 @@ const ListboxContainer = ({
     }
 
     const allowSelect = !constraints?.select && !constraints?.active;
-    const listboxOptions = {
+    const listboxOptions = extend(true, {
       __DO_NOT_USE__: {
         selectDisabled: () => !allowSelect, // can we hook this into the selections api?
+        focusSearch: isPopover,
       },
       direction: options?.direction,
-      search: disableSearch ? false : ('toggle' as stardust.SearchMode),
+      search: isPopover ? true : 'toggle' as stardust.SearchMode,
       isPopover,
-    };
+    }, dqOptionsOverrides || {});
 
     // @ts-ignore
     listboxInstance.mount(elRef.current, listboxOptions).then(() => {
@@ -92,16 +102,15 @@ const ListboxContainer = ({
   }, [listboxInstance]);
 
   return (
-    <>
-      <Box
-        height='100%'
-        border={showBorder ? '1px solid #d9d9d9' : '1px solid transparent'}
-        borderBottom={(showBorder && borderBottom) ? '1px solid #d9d9d9' : 0}
-        borderRadius='4px'
-        overflow='hidden'
-        ref={elRef}
-      />
-    </>
+    <Box
+      height='100%'
+      border={showBorder ? '1px solid #d9d9d9' : '1px solid transparent'}
+      borderBottom={(showBorder && borderBottom) ? '1px solid #d9d9d9' : 0}
+      borderRadius='4px'
+      overflow='hidden'
+      ref={elRef}
+      data-testid="filterpane-listbox-container"
+    />
   );
 };