+++ packages/sn-filter-pane/src/components/ListboxGrid/ListboxGrid.tsx
@@ -2,9 +2,6 @@ import React, {
   useCallback, useEffect, useRef, useState, useSyncExternalStore,
 } from 'react';
 import { Grid } from '@mui/material';
-import { styled } from '@mui/material/styles';
-// @ts-ignore
-import { ResizableBox } from 'react-resizable';
 import debounce from 'lodash/debounce';
 import { stardust } from '@nebula.js/stardust';
 import getWidthHeight from './get-size';
@@ -19,7 +16,6 @@ import { ExpandProps, IColumn, ISize } from './interfaces';
 import { ColumnGrid } from './grid-components/ColumnGrid';
 import { Column } from './grid-components/Column';
 import { ColumnItem } from './grid-components/ColumnItem';
-import ConditionalWrapper from './ConditionalWrapper';
 import type { IStores } from '../../store';
 import { ListboxPopoverContainer } from '../ListboxPopoverContainer';
 import useHandleActive, { ActiveOnly } from './use-handle-active';
@@ -27,11 +23,7 @@ import KEYS from '../keys';
 import { RenderTrackerService } from '../../services/render-tracker';
 import useFocusListener from '../../hooks/use-focus-listener';
 import findNextIndex from './find-next-index';
-
-// TODO: Remove
-const Resizable = styled(ResizableBox)(() => ({
-  position: 'absolute',
-}));
+import { IEnv } from '../../types/types';
 
 const prepareRenderTracker = (listboxCount: number, renderTracker?: RenderTrackerService) => {
   renderTracker?.setNumberOfListboxes(listboxCount);
@@ -43,11 +35,13 @@ const prepareRenderTracker = (listboxCount: number, renderTracker?: RenderTracke
 function ListboxGrid({ stores }: { stores: IStores }) {
   const { store, resourceStore } = stores;
   const {
-    sense, selections, keyboard, renderTracker,
+    env = {}, selections, keyboard, renderTracker,
   } = store.getState();
 
+  const { sense } = env as IEnv;
+
   // Subscribe to the resourceStore outside of react and re-render on store change.
-  const { resources } = useSyncExternalStore(resourceStore.subscribe, resourceStore.getState);
+  const { resources = [] } = useSyncExternalStore(resourceStore.subscribe, resourceStore.getState);
 
   const gridRef = useRef<HTMLDivElement>();
   const [columns, setColumns] = useState<IColumn[]>([]);
@@ -56,8 +50,12 @@ function ListboxGrid({ stores }: { stores: IStores }) {
   const { options } = stores.store.getState();
 
   const handleResize = useCallback(() => {
+    if (!resources?.length) {
+      return;
+    }
     const { width, height } = getWidthHeight(gridRef);
     const size: ISize = { width, height, dimensionCount: resources.length };
+    store.setState({ ...store.getState(), containerSize: size });
     const isSmallDevice = sense?.isSmallDevice?.() ?? false;
     const isSingleItem = resources.length === 1;
     const expandProps: ExpandProps = {
@@ -73,6 +71,8 @@ function ListboxGrid({ stores }: { stores: IStores }) {
     setColumns(expandedAndCollapsedColumns);
     prepareRenderTracker(expandedItemsCount, renderTracker);
   }, [resources]);
+
+  const dHandleResize = useRef(debounce(handleResize, isInSense ? 0 : 10));
 
   const preventDefaultBehavior = (event: React.KeyboardEvent | MouseEvent | React.MouseEvent<HTMLLIElement>) => {
     event.stopPropagation();
@@ -138,52 +138,54 @@ function ListboxGrid({ stores }: { stores: IStores }) {
   }, [keyboard]);
 
   const handleActive = useHandleActive(isInSense, selections as stardust.ObjectSelections & ActiveOnly);
-  const dHandleResize = debounce(handleResize, isInSense ? 0 : 50);
 
-  // TODO: Remove Resizable, only for developing purposes
   return (
     <>
-      <ConditionalWrapper condition={process.env.NODE_ENV === 'development'}
-        wrapper={(children: JSX.Element[]) => <Resizable width={1080} height={1000} minConstraints={[10, 10]} maxConstraints={[1220, 1820]}>{children}</Resizable>}
+      <ElementResizeListener onResize={dHandleResize.current} />
+      <Grid
+        className="filter-pane-container"
+        container
+        onKeyDown={handleKeyDown}
+        sx={{ flexDirection: isRtl ? 'row-reverse' : 'row' }}
+        columns={columns?.length}
+        ref={gridRef as unknown as () => HTMLDivElement}
+        spacing={0}
+        height='100%'
+        overflow="hidden"
       >
-        <>
-          <ElementResizeListener onResize={dHandleResize} />
-          <Grid className="filter-pane-container" container onKeyDown={handleKeyDown} sx={{ flexDirection: isRtl ? 'row-reverse' : 'row' }} columns={columns?.length} ref={gridRef as unknown as () => HTMLDivElement} spacing={0} height='100%'>
 
-            {!!columns?.length && columns?.map((column: IColumn, i: number) => (
-                <ColumnGrid key={i} widthPercent={100 / columns.length}>
-                <Column lastColumn={!isRtl ? columns.length === i + 1 : i === 0}>
+        {!!columns?.length && columns?.map((column: IColumn, i: number) => (
+          <ColumnGrid key={i} widthPercent={100 / columns.length}>
+            <Column lastColumn={!isRtl ? columns.length === i + 1 : i === 0}>
 
-                  {!!column?.items?.length && column.items.map((item: IListboxResource, j: number) => (
-                        <ColumnItem
-                          key={item.id}
-                          lastItem={column.items?.length === j + 1}
-                          listItem={item}
-                        >
-                      {item.expand
-                        ? <ListboxContainer
-                              layout={item.layout}
-                          borderBottom={(column.items?.length === j + 1) || !item.fullyExpanded}
-                              disableSearch={item.cardinal <= 3}
-                              handleActive={handleActive}
-                              stores={stores}
-                            ></ListboxContainer>
-                        : <ListboxPopoverContainer resources={[item]} stores={stores} />
-                      }
-                        </ColumnItem>
-                  ))}
+              {!!column?.items?.length && column.items.map((item: IListboxResource, j: number) => (
+                <ColumnItem
+                  key={item.id}
+                  lastItem={column.items?.length === j + 1}
+                  listItem={item}
+                >
+                  {item.expand
+                    ? <ListboxContainer
+                      layout={item.layout}
+                      model={item.model}
+                      borderBottom={(column.items?.length === j + 1) || !item.fullyExpanded}
+                      handleActive={handleActive}
+                      stores={stores}
+                    ></ListboxContainer>
+                    : <ListboxPopoverContainer resources={[item]} stores={stores} />
+                  }
+                </ColumnItem>
+              ))}
 
-                  {column.hiddenItems && overflowingResources.length
-                    && <ColumnItem height='100%'>
-                      <ListboxPopoverContainer resources={overflowingResources} stores={stores} />
-                    </ColumnItem>}
-                  </Column>
+              {column.hiddenItems && overflowingResources.length
+                && <ColumnItem height='100%'>
+                  <ListboxPopoverContainer resources={overflowingResources} stores={stores} />
+                </ColumnItem>}
+            </Column>
 
-                </ColumnGrid>
-            ))}
-          </Grid>
-        </>
-      </ConditionalWrapper>
+          </ColumnGrid>
+        ))}
+      </Grid>
     </>
   );
 }