+++ packages/sn-filter-pane/src/store/index.ts
@@ -4,9 +4,10 @@ import {
   IListBoxOptions,
   IListboxResource,
 } from '../hooks/types';
-import { ISense } from '../types/types';
+import { IEnv } from '../types/types';
 import { RenderTrackerService } from '../services/render-tracker';
 import createStore from './state-store';
+import { ISize } from '../components/ListboxGrid/interfaces';
 
 export interface IStore {
   app?: EngineAPI.IApp;
@@ -15,12 +16,14 @@ export interface IStore {
   options: IListBoxOptions;
   constraints?: stardust.Constraints;
   translator?: stardust.Translator;
-  sense?: ISense;
+  env?: IEnv;
+  directQueryEnabled?: boolean;
   embed?: stardust.Embed;
   stardustTheme?: stardust.Theme;
   selections?: stardust.ObjectSelections;
   keyboard?: stardust.Keyboard;
   renderTracker?: RenderTrackerService;
+  containerSize?: ISize;
 }
 
 interface ResourceState {