+++ types/global.d.ts
@@ -9,6 +9,11 @@ export interface IListLayout extends EngineAPI.IGenericListLayout {
     dataLayout?: 'grid' | 'singleColumn',
   },
   showTitle?: boolean,
+  qHasSoftPatches?: boolean;
+  qExtendsId?: string;
+  qMeta?: {
+    privileges?: string[];
+  }
 }
 
 // Missing properties in EngineAPI.IGenericListProperties are added in this interface until
@@ -35,4 +40,8 @@ export interface IGenericListPropertiesMissing {
       }
     }
   }
+}
+
+export interface INxAppLayoutGlobal extends EngineAPI.INxAppLayout {
+  qIsDirectQueryMode: boolean;
 }